+++ exam-com/src/main/java/com/example/entity/TOya.java
@@ -372,11 +372,17 @@ public class TOya implements IEntity {
      * @return 削除件数
      */
     public int delete() {
-        for (TEntity tEntity : this.tEntitys) {
-            tEntity.delete();
+
+        if (this.tEntitys != null) {
+            for (TEntity tEntity : this.tEntitys) {
+                tEntity.delete();
+            }
         }
-        String sql = "DELETE FROM t_oya WHERE " + getWhere();
+
+        String sql = "DELETE FROM t_entity WHERE " + getWhere();
+
         Map<String, Object> params = toMap(null, null);
+
         return Queries.regist(sql, params);
     }
